moved search logic fully to search component and improved flow of landingpage content
